Refactor SNS page with widget extraction
snsをリファクタリングしました
