+++ lib/ui/sns/view/widgets/sns_empty.dart
@@ -1,0 +1,75 @@
+import 'package:flutter/material.dart';
+
+class EmptyState extends StatelessWidget {
+  final Animation<Offset> slideAnimation;
+  const EmptyState({required this.slideAnimation, super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: SlideTransition(
+        position: slideAnimation,
+        child: Container(
+          padding: const EdgeInsets.all(40),
+          margin: const EdgeInsets.all(24),
+          decoration: BoxDecoration(
+            color: Colors.white.withOpacity(0.9),
+            borderRadius: BorderRadius.circular(25),
+            border: Border.all(
+              color: Colors.white.withOpacity(0.5),
+              width: 1,
+            ),
+            boxShadow: [
+              BoxShadow(
+                color: Colors.black.withOpacity(0.1),
+                blurRadius: 20,
+                offset: const Offset(0, 10),
+              ),
+            ],
+          ),
+          child: Column(
+            mainAxisSize: MainAxisSize.min,
+            children: [
+              Container(
+                padding: const EdgeInsets.all(20),
+                decoration: BoxDecoration(
+                  gradient: LinearGradient(
+                    colors: [
+                      const Color(0xFF667EEA).withOpacity(0.1),
+                      const Color(0xFF764BA2).withOpacity(0.1),
+                    ],
+                  ),
+                  borderRadius: BorderRadius.circular(25),
+                ),
+                child: const Icon(
+                  Icons.people_outline,
+                  color: Color(0xFF667EEA),
+                  size: 60,
+                ),
+              ),
+              const SizedBox(height: 24),
+              Text(
+                'まだフレンドがいません',
+                style: TextStyle(
+                  color: Colors.grey.shade700,
+                  fontSize: 20,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+              const SizedBox(height: 12),
+              Text(
+                'QRコードをスキャンして\nフレンドを追加しましょう！',
+                style: TextStyle(
+                  color: Colors.grey.shade600,
+                  fontSize: 14,
+                  height: 1.5,
+                ),
+                textAlign: TextAlign.center,
+              ),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
