+++ lib/ui/sns/view/widgets/sns_snackbar.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+class CustomSnackBar extends SnackBar {
+  CustomSnackBar({required String message, required bool isSuccess})
+      : super(
+          content: Row(
+            children: [
+              Container(
+                padding: const EdgeInsets.all(6),
+                decoration: BoxDecoration(
+                  color: isSuccess ? Colors.green.withOpacity(0.2) : Colors.red.withOpacity(0.2),
+                  borderRadius: BorderRadius.circular(8),
+                ),
+                child: Icon(
+                  isSuccess ? Icons.check_circle : Icons.error,
+                  color: isSuccess ? Colors.green : Colors.red,
+                  size: 20,
+                ),
+              ),
+              const SizedBox(width: 12),
+              Text(message, style: const TextStyle(fontWeight: FontWeight.w600)),
+            ],
+          ),
+          backgroundColor: Colors.white,
+          behavior: SnackBarBehavior.floating,
+          shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(15)),
+          margin: const EdgeInsets.all(16),
+        );
+}
